back_end/requirements.txt
asgiref==3.6.0
backports.zoneinfo==0.2.1
distlib==0.3.6
Django==4.1.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
filelock==3.9.1
pbr==5.11.1
platformdirs==3.1.1
sqlparse==0.4.3
testresources==2.0.1
virtualenv==20.21.0
django-cors-headers==3.14.0
Pillow
